+++ src/applications/static-pages/analytics/utilities.js
@@ -1,0 +1,3 @@
+export function isAnchor(node) {
+  return node?.nodeName === 'A';
+}
