+++ src/site/paragraphs/linkTeaser.drupal.liquid
@@ -24,12 +24,16 @@
         {% assign headerClass = "va-nav-linkslist-title" %}
     {% endif %}
 {% endif %}
-<li data-template="paragraphs/linkTeaser" data-entity-id="{{ linkTeaser.entityId }}" {% if parentFieldName === 'field_spokes' %}class="hub-page-link-list__item"{% endif %}>
+<li
+  data-template="paragraphs/linkTeaser"
+  data-entity-id="{{ linkTeaser.entityId }}"
+  data-links-list-header="{{ link.title | escape }}"
+  data-links-list-section-header="{{ section_header | escape }}"
+  {% if parentFieldName === 'field_spokes' %}class="hub-page-link-list__item"{% endif %}>
     {% assign link = linkTeaser.fieldLink | featureSingleValueFieldLink %}
-    <a
-      href="{{link.url.path}}" {% if linkTeaser.options["target"] %}target="{{ linkTeaser.options["target"] }}"{% endif %}
-      data-links-list-header="{{ link.title | escape }}"
-      data-links-list-section-header="{{ section_header | escape }}">
+    <a href="{{link.url.path}}"
+      {% if linkTeaser.options["target"] %}target="{{ linkTeaser.options["target"] }}"{% endif %}
+      >
         {% if link.title != empty %}
             <{{ header }} class="{{ headerClass }}">{{ link.title }}</{{ header }}>
             {% if parentFieldName === "field_spokes" %}
